+++ src/documents/item-datamodel-spell.mjs
@@ -1,4 +1,5 @@
 import { CastleFalkensteinItemDataModel } from "./item-datamodel.mjs";
+import { CASTLE_FALKENSTEIN } from "../config.mjs";
 
 export class CastleFalkensteinSpellDataModel extends CastleFalkensteinItemDataModel {
   static defineSchema() {
@@ -7,5 +8,9 @@ export class CastleFalkensteinSpellDataModel extends CastleFalkensteinItemDataMo
       level: new foundry.data.fields.NumberField({label: "castle-falkenstein.spell.level", initial: 0}),
       suit: new foundry.data.fields.StringField({label: "castle-falkenstein.spell.suit", initial: "?"})
     }
+  }
+
+  get suitSymbol() {
+    return CASTLE_FALKENSTEIN.cardSuitsSymbols[this.suit];
   }
 }